# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Sticky Notes Node Api Documentation",
    "version": "1.0",
    "title": "Sticky Notes Node Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Open Souce",
      "url": "https://opensource.org/licenses"
    }
  },
  "host": "sticky-notes-node-mongo.herokuapp.com",
  "basePath": "/",
...
  "paths": {
...
    "/user/{id}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "findById",
        "operationId": "findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "x-access-token",
            "description": "the JWT generated during the login process is required. Ex: abc123456.123ddgfdf.44331dfsf",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "update",
        "operationId": "update",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "User",
            "description": "User",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "in": "header",
            "name": "x-access-token",
            "description": "the JWT generated during the login process is required. Ex: abc123456.123ddgfdf.44331dfsf",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "remove",
        "operationId": "remove",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "header",
            "name": "x-access-token",
            "description": "the JWT generated during the login process is required. Ex: abc123456.123ddgfdf.44331dfsf",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    },
...
    "/auth/refresh-login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "refresh-login",
        "operationId": "refreshLogin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-access-token",
            "description": "the JWT generated during the login process is required. Ex: abc123456.123ddgfdf.44331dfsf",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/LoginResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "deprecated": false
      }
    }
  },
  "definitions": {
...
    "User": {
      "title": "User",
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": "1"
        },
        "name": {
          "type": "string",
          "example": "michael jackson"
        },
        "date_birth": {
          "type": "string",
          "format": "date-time",
          "example": "2017-05-12T13:00:05-03:00"
        },
        "description": {
          "type": "string",
          "example": "information of user"
        },
        "email": {
          "type": "string",
          "example": "email@eamil.com"
        },
        "photo": {
          "type": "string",
          "example": "image.jpg"
        },
        "password": {
          "type": "string",
          "example": "123456"
        },
        "site": {
          "type": "string",
          "example": "http://www.site.com"
        },
        "phone": {
          "type": "string",
          "example": "8197084623"
        }
      },
      "required": [
        "name",
        "date_birth",
        "description",
        "email",
        "photo",
        "password",
        "site",
        "phone"
      ]
    },
    "LoginResponse": {
      "properties": {
        "token": {
          "type": "string",
          "example": "abc123456.123ddgfdf.44331dfsf"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
  }
}